add release tags in all specs
Since we are migrating from OBS to copr, we do need to track our
packages releases now on. Also, two of the packages had missing build
dependencies and would not build on copr and two of the packages had
greater version in OBS when compared to this repository, which was also
fixed.